+++ CHANGELOG.md
@@ -3,9 +3,10 @@
 This changelog only contains the changes that are unreleased. For changes for individual releases, please visit the
 [releases](https://github.com/ATLauncher/ATLauncher/releases) page on GitHub.
 
-## 3.4.12.2
+## 3.4.13.0
 
 ### New Features
+- Remove ability to add new/relogin to Mojang accounts
 
 ### Fixes
 - Remove Vercel links from being valid from Modrinth imports